Safari-611.1.10/LayoutTests/accessibility/mac/aria-toolbar-expected.txt

item one
item tw
This tests that the aria-toolbar role correctly returns a toolbar.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS succeeded is true
PASS successfullyParsed is true

TEST COMPLETE
